Visualize Rotation

+++ Software/common/IMUData.h
@@ -1,7 +1,7 @@
 #ifndef _IMU_DATA_H_
 #define _IMU_DATA_H_
 
-#include <sys/types.h>
+#include <stdint.h>
 #include <stdio.h>
 #include <string>
 #include <iostream>
@@ -13,7 +13,7 @@ class IMUData {
 public:
 
 	typedef struct {
-		uint16_t tstamp; //timestamp
+		uint32_t tstamp; //timestamp
 		char id; //always comes out as i for imu
 		float ax; //2g: scale factor: 16384
 		float ay; //2g: scale factor: 16384
